feat: gather files into one file
